Simple refactoring, remove unnecessary socket's impl.
