+++ backend/meteo/api/src/main/java/com/meteo/models/SocketHealth.java
@@ -1,4 +1,4 @@
-package com.meteo;
+package com.meteo.models;
 
 public final class SocketHealth {
